refactor: pushRandomNumbers 함수 도메인 폴더로 이동
